Extract inventory drawers and unify split layouts
